LIBAVALON-303: Moving EnvAttribute class to utils
Also changing import modules where it's sourced

https://umd-dit.atlassian.net/browse/LIBAVALON-303
